+++ lektor/admin/modules/livereload.py
@@ -1,5 +1,9 @@
+from __future__ import annotations
+
 import queue
 import secrets
+from typing import Generator
+from typing import TYPE_CHECKING
 
 from flask import Blueprint
 from flask import render_template
@@ -7,6 +11,10 @@ from flask import url_for
 
 from lektor.admin.utils import eventstream
 from lektor.reporter import reporter
+
+if TYPE_CHECKING:
+    from flask.typing import ResponseReturnValue
+    from lektor.builder import Artifact
 
 PING_DELAY = 1.0
 
@@ -17,21 +25,21 @@ version_id = secrets.token_urlsafe(16)
 
 @bp.route("/events")
 @eventstream
-def events():
-    events = queue.Queue()
-    with reporter.on_build_change(events.put):
+def events() -> Generator[dict[str, str], None, None]:
+    updated_artifacts: queue.Queue[Artifact] = queue.Queue()
+    with reporter.on_build_change(updated_artifacts.put):
         while True:
             yield {"type": "ping", "versionId": version_id}
 
             try:
-                change = events.get(timeout=PING_DELAY)
+                artifact = updated_artifacts.get(timeout=PING_DELAY)
             except queue.Empty:
                 continue
-            yield {"type": "reload", "path": change.artifact_name}
+            yield {"type": "reload", "path": artifact.artifact_name}
 
 
 @bp.route("/events/worker.js")
-def worker_script():
+def worker_script() -> ResponseReturnValue:
     return (
         render_template("livereload-worker.js", events_url=url_for(".events")),
         200,